requirements.txt
streamlit==1.51.0
ultralytics==8.3.227
opencv-python-headless==4.12.0.88
torch==2.9.0
torchvision==0.24.0
Pillow==12.0.0
